device config

+++ app/src/main/java/com/SmartWatchVoice/bestapp/MainActivity.java
@@ -132,7 +132,7 @@ public class MainActivity extends AppCompatActivity {
 
     private void onSpeechExpect(JSONObject action, Object extra, OnResultCallback callback) throws JSONException {
         JSONObject payload = action.getJSONObject("payload");
-        int timeout = payload.getInt("timeout");
+        Long timeout = Long.valueOf(payload.getInt("timeout"));
         Utils.sendToHandlerMessage(RuntimeInfo.getInstance().speechFragmentHandler, HandlerConst.MSG_EXPECT_SPEECH, timeout);
 
         String dialogId = payload.getString("dialogId");
